feat: resolving react query invalidatequery issues and refactoring
