+++ app/(tabs)/index.tsx
@@ -3,23 +3,21 @@ import { View, StyleSheet } from 'react-native';
 import { Button } from '@rneui/themed';
 import { useState } from 'react';
 import { AddWorkoutFormOverlay } from '@/components/AddWorkoutForm';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { QueryClient, QueryClientProvider, useQueryClient } from '@tanstack/react-query';
 
-const queryClient = new QueryClient();
+export const queryClient = new QueryClient();
 
 export default function Tab() {
-  const [visible, setVisible] = useState<boolean>(false);
+  // const [visible, setVisible] = useState<boolean>(false);
 
-  const handleToggleOverlay = () => {
-    setVisible((prev) => !prev);
-  };
+  // const handleToggleOverlay = () => {
+  //   setVisible((prev) => !prev);
+  // };
 
   return (
     <QueryClientProvider client={queryClient}>
       <View style={styles.container}>
         <Workouts />
-        <Button title="Add Workout" onPress={handleToggleOverlay} />
-        <AddWorkoutFormOverlay visible={visible} handleToggleOverlay={handleToggleOverlay} />
       </View>
     </QueryClientProvider>
   );
