+++ hooks/query.ts
@@ -1,0 +1,9 @@
+import { queryClient } from '@/app/(tabs)';
+
+export const workoutsQueryKey = 'workouts';
+
+export const invalidateWorkoutsQuery = async () => {
+  await queryClient.invalidateQueries({
+    queryKey: [workoutsQueryKey],
+  });
+};
